SamplingPriority gets its own component
