+++ BUILD.bazel
@@ -15,6 +15,8 @@ cc_library(
         "src/propagation.h",
         "src/sample.cpp",
         "src/sample.h",
+        "src/sampling_priority.cpp",
+        "src/sampling_priority.h",
         "src/span.cpp",
         "src/span.h",
         "src/span_buffer.cpp",
